Raise the number of interrupt sources per CPU from 32 to 64.  This effectively triples the number of interrupt vectors that can be handled by the primary CPU.  Important for MSI, but could also fix some issues with large machines loaded with a lot of devices.
tested by many; ok deraadt@, marco@

+++ sys/arch/amd64/include/cpu.h
@@ -1,4 +1,4 @@
-/*	$OpenBSD: cpu.h,v 1.66 2011/04/13 02:49:12 guenther Exp $	*/
+/*	$OpenBSD: cpu.h,v 1.67 2011/05/23 09:52:24 claudio Exp $	*/
 /*	$NetBSD: cpu.h,v 1.1 2003/04/26 18:39:39 fvdl Exp $	*/
 
 /*-
@@ -82,11 +82,11 @@ struct cpu_info {
 	struct pcb *ci_idle_pcb;
 
 	struct intrsource *ci_isources[MAX_INTR_SOURCES];
-	u_int32_t	ci_ipending;
+	u_int64_t	ci_ipending;
 	int		ci_ilevel;
 	int		ci_idepth;
-	u_int32_t	ci_imask[NIPL];
-	u_int32_t	ci_iunmask[NIPL];
+	u_int64_t	ci_imask[NIPL];
+	u_int64_t	ci_iunmask[NIPL];
 #ifdef DIAGNOSTIC
 	int		ci_mutex_level;
 #endif
